requirements.txt
 streamlit as st
 numpy as np
 pandas as pd
 matplotlib.pyplot as plt
 seaborn as sns